toolkit/content/tests/chrome/frame_subframe_origin_subframe1.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>

<page id="frame1"
      style="background-color:green;" 
      xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>

<spacer height="10px"/>
<iframe
  style="margin:10px; min-height:170px; max-width:200px; max-height:200px; border:solid 1px white;"
  src="frame_subframe_origin_subframe2.xul"></iframe>
<spacer height="3px"/>
<caption id="cap1" style="min-width:200px; max-width:200px; background-color:white;" label=""/>
<script class="testbody" type="application/javascript">

// Fire a mouse move event aimed at this window, and check to be
// sure the client coords translate from widget to the dom correctly.

function runTests()
{
  synthesizeMouse(document.getElementById("frame1"), 3, 4, { type: "mousemove" });
}

function mouseMove(e) {
  e.stopPropagation();
  var element = e.target;
  var el = document.getElementById("cap1");
  el.label = "client: (" + e.clientX + "," + e.clientY + ")";
  parent.opener.wrappedJSObject.SimpleTest.is(e.clientX, 3, "mouse event clientX on sub frame 1");
  parent.opener.wrappedJSObject.SimpleTest.is(e.clientY, 4, "mouse event clientY on sub frame 1");
  // fire the next test on the sub frame
  frames[0].runTests();
}

window.addEventListener("mousemove", mouseMove, false);

</script>
</page>
